Update README.md
This release introduces a popup reminder for the destruction of the Beis Hamikdash that appears on all websites. The popup can be easily dismissed by clicking on it. Minor improvements were made to ensure smoother integration with Tampermonkey, and instructions have been added to the README for easier installation and use.